requirements.txt
feedparser>=6.0.10
requests>=2.31.0
anthropic>=0.18.0
slack-bolt>=1.18.0
slack-sdk>=3.23.0
python-dotenv>=1.0.0
APScheduler>=3.10.4
pytz>=2023.3
beautifulsoup4>=4.12.2
thefuzz>=0.19.0
python-Levenshtein>=0.21.1

